Consolidate Contents component and context with ContentsList
