+++ src/components/Chapters/RightColumn.js
@@ -5,8 +5,8 @@ import LeafyGreenCard from '@leafygreen-ui/card';
 import { css, cx } from '@leafygreen-ui/emotion';
 import Icon from '@leafygreen-ui/icon';
 import { uiColors } from '@leafygreen-ui/palette';
-import ContentsList from '../ContentsList/ContentsList';
-import ContentsListItem from '../ContentsList/ContentsListItem';
+import ContentsList from '../Contents/ContentsList';
+import ContentsListItem from '../Contents/ContentsListItem';
 import Link from '../Link';
 import { SidenavContext } from '../Sidenav';
 import useActiveHeading from '../../hooks/useActiveHeading';
