+++ src/components/Contents/Contents.js
@@ -2,10 +2,10 @@ import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { ContentsContext } from './contents-context';
-import ContentsList from './ContentsList/ContentsList';
-import ContentsListItem from './ContentsList/ContentsListItem';
-import { theme } from '../theme/docsTheme';
-import { formatText } from '../utils/format-text';
+import ContentsList from './ContentsList';
+import ContentsListItem from './ContentsListItem';
+import { theme } from '../../theme/docsTheme';
+import { formatText } from '../../utils/format-text';
 
 const StyledContents = styled('div')`
   @media ${theme.screenSize.largeAndUp} {
