+++ src/components/Heading.js
@@ -8,7 +8,7 @@ import useScreenSize from '../hooks/useScreenSize';
 import TabSelectors from './TabSelectors';
 import { TabContext } from './tab-context';
 import ConditionalWrapper from './ConditionalWrapper';
-import Contents from './Contents';
+import Contents from './Contents/Contents';
 import Permalink from './Permalink';
 
 const FeedbackHeading = Loadable(() => import('./Widgets/FeedbackWidget/FeedbackHeading'));
