+++ src/components/Sidenav/GuidesTOCTree.js
@@ -4,7 +4,7 @@ import { cx } from '@leafygreen-ui/emotion';
 import { SideNavItem } from '@leafygreen-ui/side-nav';
 import { sideNavItemTOCStyling } from './styles/sideNavItem';
 import Link from '../Link';
-import { ContentsContext } from '../contents-context';
+import { ContentsContext } from '../Contents/contents-context';
 import { formatText } from '../../utils/format-text';
 import { getPlaintext } from '../../utils/get-plaintext';
 import { isCurrentPage } from '../../utils/is-current-page';
